fix messages (#5421)
